Issue:
In CPL, during a proxy to a location set the
value of the branch flag being passed to append_branch
is currently the configured value for "nat_flag" (ie 6).
However, append_branch is expecting a bit mapped flag value (ie 0x40).

Also in CPL, the nat flag is getting set for the msg
if any branches have the nat flag set. This can cause
the RURI branch to have the nat flag set when the
location associated with the RURI branch does not
have the branch flag set.

Also, to be consistant with the default used in the
registrar module, made the default for nat flag -1
in the cpl-c module.

Patch provided by John Riordan <john@junctionnetworks.com>


git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@1229 689a6050-402a-0410-94f2-e92a70836424