This patch to uri_radius module adds Acct-Session-Id to Access-Request According to RFC2866.
It is similar to patch 2046495 that does the same for the auth_radius
module, and to patch 2049305 for the avp_radius module, and to patch
2051073 for the group_radius module .

From RFC2866:
"
5.5. Acct-Session-Id
Description
This attribute is a unique Accounting ID to make it easy to match start and
stop records in a log file. The start and stop records for a given session
MUST have the same Acct-Session-Id. An Accounting-Request packet MUST have
an Acct-Session-Id. An Access-Request packet MAY have an Acct-Session-Id;
if it does, then the NAS MUST use the same Acct-Session-Id in the
Accounting- Request packets for that session.
"

The purpose of this patch is to create a relationship between an auth and
acct request, based on SIP CALL-ID value (Acct-Session-Id attribute).


Closes patch #2086267.
Credits goto Dror Wald.


git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@4759 689a6050-402a-0410-94f2-e92a70836424