Backport from trunk (rev #5324)
At charset_test() in opensipsdbctl.mysql script. grep is called with an
option which is not supported by standard grep in Solaris.
This patch proposes a different approach for that case; it defines a
variable which counts how many times user has failed in passing a supported
charset; after 3 tries, script exits with error code of 1.

Closes 2577793.
Credits go to Sergio Gutierrez.


git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/branches/1.4@5325 689a6050-402a-0410-94f2-e92a70836424