Rollback changes made in revision #5804
The checks are unnecessary and they only hide the error by no longer sending
keepalive messages to the affected endpoints. In reality contact->uri should 
always be sip:IP:port so no extra checks are required. The port is always
present even when default and the uri is built from the msg->rcv structure
so it doesn't depend on any user setting from the script.

In the end the segfault cause has proven to be in the sdp parser code that
corrupted the shared memory.



git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/branches/1.5@5910 689a6050-402a-0410-94f2-e92a70836424