Several in-DIALOG PUBLISH fixes.
The number of generated PUBLISH requests is no optimal.

Disabled PUBLISHing for in-dialog REQUESTS. PUBLISHing (again) the
"confirmed" state when handling OPTIONS, ACK, BYE, and other dialog related
requests triggers a mess of useless NOTIFY requests. We only keep PUBLISHing
on reINVITE/UPDATEs as that might be required to be compatible with session
timers (they refresh the PUBLISH lifetime).

Also added a flag to allow inhibiting PUBLISHing for INVITE/UPDATE in-dialog
requests. That might be required for "things" like Asterisk reINVITEs for which
we do not want rePUBLISHing again our "confirmed" state. Especially useful when
you know that Asterisk reINVITEs for INFO, BYE, and so on.