Replace odd -j in module-submake with ability for proper parallelism.
Recently `-j` was added to the list of flags passed to the modules
$(MAKE) command. That's bad practise for two reasons: (1) the caller
should decide on how much -j is required, not the programmer, and (2)
just -j without any numeric argument is "use all resources you can
get", which can be too much.

We order the Makefile around so the supermake (the user invocation)
can dispatch individual submakes in a parallel fashion. This is done
by (a) making the module directories phony targets and (b) replacing
the sequential for-loop over the modules with a simple dependency.