digest+m_auth: Allow a domain in the digest username.
Before this patch, a domain in the digest username was allowed:

    Digest username="abc@domain", realm="domain"

but only if the domain is equal to the realm. This was introduced in
a92bf789 (Juha Heinanen, Mar. 29 2006).

The SIP spec. doesn't mention any such restriction.

This patch undoes that commit and makes sure the whole username is used
when authenticating through `pv_www_authorize` and `pv_proxy_authorize`.
That makes the following valid:

    Digest username="abc@domain", realm="something different"