Added 2 stage update for UAS side of a transaction.
The cloned UAS msg (from a transaction) can be now updated (new_uri, dst_uri, lumps, etc) after the orignal cloning. Also, a original cloning can be done without saving the updatable part of the UAS msg (like when you do t_newtran() in script, a partial UAS is cloned and full cloning (update) is done on t_relay() ) - this allows the transaction to remember all the changes you do between the creation of the transaction and the actual relay.