Fix balancing on failover.
First, the way the calls were re-distributed if the calculated destination is inactive was completely unbalanced and did not respect the hashing concepts (same hash hits the same destination). Along with this problem, the fix covers the re-balancing problem - if a destination becomes inactive, the balancing over the remaining destination should not be affected (if hash H1 hits destination D1, after D2 goes inactive, H1 must still hit D1).
The use_default concept was reworked to properly work with all algorithms and in all cases (with or without weights).

Closes #387.