s.fill transformation: Prioritize string pvars over integer ones
Reason for this is that a pvar which holds a string is a more
particular occurrence than an int one, since integers can be
easily stringified.

This fixes some issues with pvars which have multiple internal
representations (e.g. $rm is "INVITE" and 1 at the same time)

Reported by Eric Tamme