Fix restoring/freeing the path_vec in failure route.
The path_vec (array holding the PATH string) was not properly pushed into the faked requests before calling the failure route. This was generating a mixture of shm/pkg memory when the faked request had to be freed.
This crash was trigger only if PATH support was used and requests were looked up with PATH header.

Reported by @AVFedorov and @petekelly.
Closes #484

Many thanks to @AVFedorov for his details information and to @petekelly for support in investigating and testing.