Add CRL (Certificate Revocation List) verification for TLS
CRL verification was added in the proto tls module. By adding this
feature revoked client certificates can be detected during the TLS
process, thus permitting a robust security mechanism. The following
parameters where added to the proto tls module in the configuration
script: crl_dir and crl_check_all. The crl_dir parameter specifies the
directory which contains the CRL files (multiple CRL files can be
added).
The crl_check_all parameter must be 0 or 1 and specifies whether all the
certificates from the chain are verified against a CRL or not.
By default, CRL verification is enabled only for client certificates
(or when crl_check_all is 0). If crl_check_all is 1 then the
issuer (chain) certificates are also verified against the given CRL
files.