b2b_logic: set late SDP negociation flag
For the entity that initiates a call, late SDP negociation is used, because
the SDP of the first entity is not known. This patch sets the late_sdp flag
for that entity, to flag the b2b logic to send the ACK with SDP.

This bug was reported on the mailing list by Sebastian Sastre.