Sipcapture: store in the database without passing through script
	With this new feature it is possible to store the HEP
message inside the database without going through the script,thus
allowing you to use the script for processing valid messages,
since the ones that are send to this node aren't intended for it.
	There have also been defined a return code from the hep
callbacks(HEP_SCRIPT_SKIP) which allows the type of behaviour
explained earlier. If at least one HEP callback sets this return
code then passing through the script shall be skipped.