Fix race condition between retransmission and wait timer in TM.
Overlapping of retransmission timer and wait timer may lead to inconsistent timer list -> various crashes around TM timer lists.
Related to #749, #732, #723.
Many thanks for Trevor Francis ( @46labs )  for helping with the troubleshooting.

IMPORTANT NOTE: this is a temporary fix as it simply avoid the collision between the two timer routines - normally the routines themselves should be safe to run in parallel (more digging is required here).