Fixed autodetection of the reactor size (open files)
If we auto-detect the size of the reactor(via rulimit), be sure we have enough pkg memory (and do the necessary correction).
If the open files limit is given, warn if the reactor will use too much memory.
Thanks to MAxim Sobolev for reporting it via mailing list and to @cepehutu
Closes #778