add ASYNC_CHANGE_FD status in reactor
Allows changing the file descriptor in the resume function. Such
functionality is useful for failover purposes, like having
multiple destinations, you find that one is not working only in
the resume function and you want to try another one. Now you
can change the file descriptor, remove the old one and add the
new descriptor and then wait for it to be triggered.