hold db state between async call and resume function
Databases can now keep their state between the async call and the
resume parameter using a newly added (void*) parameter. It is the
job of the upper layer(the module calling the async function) to
store this database state inside the (void*) parameter they are using
to hold their state (see avpops and sipcapture async handling).

Also this patch handles the new ASYNC_CHANGE_FD state in avpops
and sipcapture.