t_newtran() automatically sends a 100 Tryting for INVITE.
TO have a consistent behavior between:
	t_newtran();
	t_relay();
and
	t_relay();

(where in the first case a 100 Trying will not be auto generated), we do not sent the 100 Trying on r_relay (when transaction is built) but on t_newtan().
Reported by MAxim Sobolev.

(cherry picked from commit 0db77896288e7ef093352fc0a08f4035c1191bf4)