Fixed memory leak on checking disposition hdr.
Free the parsed disposition body all the time (on success and failure) as it is not attached to any structure, so it will leak if not explicitly be freed.
Credits for reporting this go to Jarrod Baumann.
Based on an original fix from Jarrod Baumann
Closes PR #825

(cherry picked from commit 4601e5653be202912a3732f08cdafc63d64afe7f)