Revert commit 1e1ea14
Do not automatically set db_url to the db_default_url as we have no idea if the default partition is gone be used. By forcing it we create some extra unwanted dependencies (like setting the version for the defaul table, even if not used).
Even if we break the backward compatibility, makes more sens to explicitly set the db_url if you use the default partition