[acc] fix double free issue
Since callbacks were registered only at first, and aftewards only
flags were changed there was no way to know if tm is to free the
shmed flags from acc or dialog. For this a new flag was introduced
to indicate whether the flags shall be removed either by dialog or
tm.