TCP connection lookup: Add matching by transport
The fact that all currently supported TCP-oriented protocols have different
standard ports is not an excuse for not including the transport into the
match condition (i.e. dest IP address + dest port + dest transport).

This may very well prevent issues with messages being sent out over the
wrong TCP connection in a number of scenarios where we employ TCP connection
reusage over both plain TCP and TLS with a Proxy that specifies the same port
in the topmost Via header when it establishes each of the two connections.