Fix memory corruption on $stat() vars
The $stat() var specs are kept in pkg, while the name (if stat not found) is kept in memeory. This means all the copies (form all procs) of the specs will point to a single shm block. It is not safet to free this block as you have no idea how many copies of the spec (from other processes) still refer the name.
Closes #902

(cherry picked from commit 4baead765179f6bfbc098b3e59791d7bbc4a8d42)