Fix socket selection for outbound UDP.
Do not use the bind_address of the process when comes to send an UDP package (bind_address may be different from process to process, so the selection may give different results, depending on the processes sending the UDP).
If you have a sip_msg, try using received interface of the msg (if matches), otherwise use the first UDP interface.