clusterer: complete module rework
	The clusterer module was completely reworked, starting with code refactoring and coding style improvements,
in order to introduce better node availability detection and inter-cluster, failover message routing mechanisms.
	As there were no explicit probes at the clusterer module level, an active pinging mechanism was introduced
which establishes the state of the connections of the current node with other nodes in the cluster. As such,
when a node is unresponsive, it is not automatically considered down but only the direct 'link' with that node
will be marked as failed. The current node will have to take into consideration the perspective of the other
nodes in the cluster in respect to the state of the presumably failed node. This leads to the need of having
an awareness of the entire topology in the cluster and when every other node has no link with a specific node,
that node is reported as down.
	A link-state resembling protocol was implemented in order to learn the cluster topology and compute
a routing 'table' used for sending bin messages between nodes.
	The clusterer module API was improved and its usage in the dialog, usrloc and ratelimit modules was adapted.