clusterer: improve locking mechanism
Replaced the lock that protected any access to the shm structures with a RW lock that protects the cluster list and corresponding node lists for insertions/deletions. Also, locks were added to the cluster info and node info structures that protect accesses to structure fields that may be changed when the RW lock is acquired for reading. In consequence potentially blocking operations, like message sending, can now be done without holding any lock(except the RW lock acquired for reading, which is safer because most of the time it won't block).