[proto_hep] fix bug introduced in 7190a2d
	Commit 7190a2d was freeing the extra chunks twice when
the message was processed as SIP. Now the original hep context
free function is not registered as a context free function, but
called manually each time all the hep processing is done. This
way we avoid the memory leaks that appeared when the message was
not a SIP message, and the chunks were not freed.