Fix consistency and use dialog_id.
Dialog_id is printed by dlg_list (along the h_label and h_id) as a simpler alternative to identify the dialog.
Also, the dlg_end_dlg accepts both h_label and h_id or a dialog_id.
This is needed as the DB exposes only the dialog_id, and the dialogs from DB cannot be correlated anymore with the MI commands. See #963

Closes #963.