Fix CANCELLing of transactions not forwarded.
Due async operations, a transaction may be created and it may exist for a long time before any branch is created for it (via a t_relay). The current fix takes care and does proper handling on incoming CANCELs matching the transaction while it has no branch (no t_relay() on it).
Old behavior - the INVITE transaction was never getting a final reply
New brhavior - the INVITE transaction gets an internally generated 487.