Added content-aware body (multi-part) changing.
The new code allows parsing and modification of the body via the "body parts" list - parts can be modified (with function or lumps), can be deleted and new ones can be added. Any type of modification (and any of them) is allowed.
The body is completely rebuild based on the resulting list of body parts.
Also the new code supports transparent transitions from no body to body, single part body to multi-part body or the other way around - with automatic update of the Content-Type and Content-Len headers, of course.