Fix recalculating Active nodes and weight sums after reload
When inheriting the state from in-mem old data set, the state of a node may change, so we need to recalculate the number of active nodes and the running sums (for weights) in the set.

Fixes #972, but we need @hydrosine confirmation on this.

(cherry picked from commit c5c1f40ea54f0908421ece09bdc5e4b70ceba6cb)