rest_client: Fix issues with libcurl parallel transfers
We cannot use the "parallel transfers" feature of libcurl's multi interface
because that would consume read events from some its file descriptors that
had previously been added to the OpenSIPS reactor. This would sometimes lead
to dangling descriptors in the reactor, as well as some OpenSIPS async routes
which are not triggered.

To work around this, we can still achieve the desired effect with a pool of
multi handles each doing a single transfer, rather than using 1 multi handle
doing multiple transfers.

Thanks for reporting and troubleshooting to @arovetto and Agalya Ramachandran

Fixes issue #997