Added support for cloning (shm and pkg) for the sip body structure.
This is needed to properly implement per-branch changes of the msg bodies and for failure route body restore.
This commit adds functions for the cloning the sip_msg_body and body_part structures (into shm or pkg, in the same or different sip_msg structure). It also provides support to allow body part parsers (SDP, ISUP, etc) to do their own cloning stuff.