drouting: Fix out-of-bound initialization
Upon load/reload of drouting tables, if parse_destination_list() needs to
increase destinations array size, it asks for a new array twice the size of
the previous one.
It memset() the newly allocated capacity, skipping memory that is to be filled
with already learnt destinations, but using the array size instead of the
remaining capacity.
This patch fixes the size for this memset().

Closes OpenSIPS/opensips#1039

(cherry picked from commit b11ce9723f1ca43cb7646b96e51dfa306a6280d7)