Added 'launch' statement.
'launch' starts a job in async mode and continue the script immediatly, without waiting for the job to be terminated. A "report" route may be triggered (outside the context of the msg launching the job) when the job is terminated to be able to access the return code and the possible returned data.