Bind proto to all interfaces if no listeners declared
OpenSIPS 1.11 used to bind to each available interface if no 'listen'
directives declared in configuration file. This was rather convenient
when there are multiple network interaces and we want to bind at all
of them. But after SIP transport protocols rework this feature has
gone. So the main purpose of this change is restoring this feature.
How it works now: if some transport module is loaded via 'loadmodule'
statement in configuration file and no listeners are declared for this
module via 'listen' statement then listeners list will be populated
with all network interfaces that are currently available.