Set 'CURLOPT_FAILONERROR' option to '0' in rest_client module
This is done in order to archive following goals:
- suppress error log messages like 'curl_easy_perform: ...' when HTTP
  server returns response with status code larger or equal to 400.
- make 'rest_xxx' functions to have the same behaviour with different
  Curl distributions. Now for different Curls these functions may
  return different exit codes for the same HTTP response.

There is an option in Curl named 'CURLOPT_FAILONERROR'. If it's set to
'1' then if returned HTTP code is equal to or larger than 400 the HTTP
request is considered as failed and 'curl_easy_perform' returns non
'CURLE_OK' code. In vanilla (upstream) Curl this option is set to '0'
(do not fail) but some patched Curl versions set this option to '1'
(at least this is true for Debian Wheezy). As result we have a
different behaviour for different Curl derivatives. So it seems to be
better to enforce setting this option to '0' in order to have the same
behaviour for different Curls.

(cherry picked from commit 7bb059fa9cbfe6222d7eb7ebad4a2d51c4a25a8b)

Conflicts:
	modules/rest_client/rest_methods.c