Handle unset DB_PATH in opensipsctl.sqlite
When DB_PATH is unset, sqlite_query will invoke the sqlite executable
with a single parameter. This causes sqlite to interpret the query as
the path to the database file and subsquently enter interactive mode,
causing opensipsctl to freeze. Prevent this by setting DB_PATH to a
default value when the variable is unset.