tls_mgm: properly get the domain name and address from the corresponing DB columns
Previously, the value of the domain column was only an id and the actual domain name
was taken from the address column when it's value was not in the "ip:port" format.
This was somewhat misleading and consequently the domain column now expects the
domain name, and the address column an ip address and port.
Also drop the not null constraint from the domain and address DB columns, as server
domains only need and address and client domains can be either socket or name based.