[siptrace] set context for all failure route requests
	The siptrace context pointer was handled via the processing
context. In failure route the context was not set thus not knowing
if the tracing was activated or not. Because of this internally
generated requests were not traced. Now each time a request is
sent the context is set from tm callback parameter.

(cherry picked from commit 086b3caf8b87e4916bbfca134c2d119d3ee78e5c)