Fixed updating transaction if no relay done in request route.
If the request route ends without any t_relay() (like in the PN scenario, where there are no contact registered at the time of INVITE handling), we need to force upating the cloned request (in the UAS side of the transaction) with all the changes done over the request in script (normally this is done at the t_relay() time).

Reported by Nick Altmann.