Remove bogus auto conversion from SIP uri to TEL uri
For ancient reasons, a SIP URI with user=phone was automatically converted to a TEL URI. Such conversion, automatically done, is dangerous - there is nothing in the RFC3261 stating something like this. Even more, the conversion is not complete - besides moving the username parameters to URI parameters, the domain is not stripped and the TEL not added.

(cherry picked from commit 05ece794083f8328cbd5259cdb7b3aead8d9f943)