mid_registrar: Fix mem corruption on SIP retransmissions
The TMCB_RESPONSE_IN function gets called once per each retransmission,
and we want to avoid this in the mid_registrar code. Since the callbacks
themselves are mutually exclusive, we can get away with zeroizing our own
parameter after the first run.

A better way of fixing this would be to extend the tm callback API to
include SIP retransmission preferences.