protos/tcp: respect async operations order
In case there is a write request, before actually writing the data on
the network, check if there are any async chunks left, and make sure
they are sent before any other write happens. This is necessary to make
sure the order of the packets is respected, otherwise we might end up
with fragments of other packets in the stream.

Thanks go to Vlad Patrascu for reporting and testing this.

(cherry picked from commit 5773e65b9058643191544763e0f5ff26f5a7fb3b)